requirements.txt
youtube-dl
python-telegram-bot
proxybroker
tenacity
quart
pycountry
